chore: TrackFormContainer replace route param with query param
